requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.31.0
cloudinary-upload==0.0.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.1.6
django-autoslug==1.9.8
django-cloudinary-storage==0.3.0
django-environ==0.9.0
django-js-asset==2.0.0
django-mdeditor==0.1.20
django-storages==1.13.2
fonttools==4.39.0
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
pyotp==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
qrcode==7.4.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
